requirements.txt
pandas
numpy
lightgbm
scikit-learn
scipy
sqlalchemy
python-decouple
optuna